+++ torchrl/collectors/distributed/__init__.py
@@ -2,157 +2,8 @@
 #
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
-import subprocess
-import time
 
-import torchrl
-
-from . import DistributedDataCollector
-from ...data import BoundedTensorSpec
-from ...envs import EnvCreator
-from ...envs.libs.gym import GymEnv
-from ..collectors import RandomPolicy
-from .generic import (
-    _distributed_init_delayed,
-    DEFAULT_SLURM_CONF,
-    DEFAULT_SLURM_CONF_MAIN,
-    DistributedDataCollector,
-)
-from .rpc import _rpc_init_collection_node, RPCDataCollector
+from .generic import DEFAULT_SLURM_CONF, DistributedDataCollector
+from .rpc import RPCDataCollector
 from .sync import DistributedSyncDataCollector
-
-
-SUBMITIT_ERR = None
-try:
-    import submitit
-
-    _has_submitit = True
-except ModuleNotFoundError as err:
-    _has_submitit = False
-    SUBMITIT_ERR = err
-
-
-class submitit_delayed_launcher:
-    """Delayed launcher for submitit.
-
-    In some cases, launched jobs cannot spawn other jobs on their own and this
-    can only be done at the jump-host level.
-
-    In these cases, the :func:`submitit_delayed_launcher` can be used to
-    pre-launch collector nodes that will wait for the main worker to provide
-    the launching instruction.
-
-    Args:
-        num_jobs (int): the number of collection jobs to be launched.
-        framework (str, optional): the framework to use. Can be either ``"distributed"``
-            or ``"rpc"``. ``"distributed"`` requires a :class:`~.DistributedDataCollector`
-            collector whereas ``"rpc"`` requires a :class:`RPCDataCollector`.
-            Defaults to ``"distributed"``.
-        backend (str, optional): torch.distributed backend in case ``framework``
-            points to ``"distributed"``.
-            Defaults to ``'gloo'``.
-        tcpport (int or str, optional): the TCP port to use. Defaults to ``10003``.
-        submitit_collection_conf (dict, optional): the configuration to be passed to submitit.
-            Defaults to :obj:`torchrl.collectors.distributed.generic.DEFAULT_SLURM_CONF`
-
-    Examples:
-        >>> num_jobs=2
-        >>> @submitit_delayed_launcher(num_jobs=num_jobs)
-        ... def main():
-        ...     from torchrl.envs.libs.gym import GymEnv
-        ...     from torchrl.collectors.collectors import RandomPolicy
-        ...     from torchrl.data import BoundedTensorSpec
-        ...     collector = DistributedDataCollector(
-        ...         [EnvCreator(lambda: GymEnv("Pendulum-v1"))] * num_jobs,
-        ...         policy=RandomPolicy(BoundedTensorSpec(-1, 1, shape=(1,))),
-        ...         launcher="submitit_delayed",
-        ...     )
-        ...     for data in collector:
-        ...         print(data)
-        ...
-        >>> if __name__ == "__main__":
-        ...     main()
-        ...
-    """
-
-    def __init__(
-        self,
-        num_jobs,
-        framework="distributed",
-        backend="gloo",
-        tcpport=10003,
-        submitit_main_conf: dict = DEFAULT_SLURM_CONF_MAIN,
-        submitit_collection_conf: dict = DEFAULT_SLURM_CONF,
-    ):
-        self.num_jobs = num_jobs
-        self.backend = backend
-        self.framework = framework
-        self.submitit_collection_conf = submitit_collection_conf
-        self.submitit_main_conf = submitit_main_conf
-        self.tcpport = tcpport
-
-    def __call__(self, main_func):
-        def exec_fun():
-            if not _has_submitit:
-                raise ModuleNotFoundError(
-                    "Failed to import submitit. Check installation of the library."
-                ) from SUBMITIT_ERR
-            # submit main
-            executor = submitit.AutoExecutor(folder="log_test")
-            executor.update_parameters(**self.submitit_main_conf)
-            main_job = executor.submit(main_func)
-            # listen to output file looking for IP address
-            print(f"job id: {main_job.job_id}")
-            time.sleep(2.0)
-            node = None
-            while not node:
-                cmd = f"squeue -j {main_job.job_id} -o %N | tail -1"
-                node = subprocess.check_output(cmd, shell=True, text=True).strip()
-                try:
-                    node = int(node)
-                except ValueError:
-                    time.sleep(0.5)
-                    continue
-            print(f"node: {node}")
-            cmd = f"sinfo -n {node} -O nodeaddr | tail -1"
-            rank0_ip = subprocess.check_output(cmd, shell=True, text=True).strip()
-            print(f"IP: {rank0_ip}")
-            world_size = self.num_jobs + 1
-
-            # submit jobs
-            executor = submitit.AutoExecutor(folder="log_test")
-            executor.update_parameters(**self.submitit_collection_conf)
-            jobs = []
-            if self.framework == "rpc":
-                from .rpc import DEFAULT_TENSORPIPE_OPTIONS
-
-                tensorpipe_options = DEFAULT_TENSORPIPE_OPTIONS
-            for i in range(self.num_jobs):
-                rank = i + 1
-                if self.framework == "distributed":
-                    job = executor.submit(
-                        _distributed_init_delayed,
-                        rank,
-                        self.backend,
-                        rank0_ip,
-                        self.tcpport,
-                        world_size,
-                    )
-                elif self.framework == "rpc":
-                    job = executor.submit(
-                        _rpc_init_collection_node,
-                        rank,
-                        rank0_ip,
-                        self.tcpport,
-                        world_size,
-                        None,
-                        tensorpipe_options,
-                    )
-                else:
-                    raise NotImplementedError(f"Unknown framework {self.framework}.")
-                jobs.append(job)
-            for job in jobs:
-                job.result()
-            main_job.result()
-
-        return exec_fun
+from .utils import submitit_delayed_launcher
